requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-yasg==1.21.7
apscheduler==3.10.1
cron-validator==1.0.3
python-dateutil==2.8.2
gunicorn==21.2.0
django-redis==5.4.0